Implemented Industry View & (temp)Project ViewUI
Implemented simple industry view and project view.
Now can view buildingds and build more workplaces. Just the bare min, so need to implement ownership of the building, costs of the building, hiring builders, and etc later on.
Minor bug fix to change happiness where it was checking health for out
of bonds check.